requirements.txt
# Backend dependencies
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
requests==2.31.0
openai==1.3.7
langgraph==0.0.62
langchain==0.0.350
langchain-openai==0.0.2
pydantic==2.5.0
python-multipart==0.0.6
aiofiles==0.23.0

# For CORS
fastapi-cors==0.0.6

# For HuggingFace Spaces
gradio==4.12.0

# Additional dependencies
pathlib2==2.3.7
